+++ R/try_check.R
@@ -10,15 +10,9 @@
 try_check <- function(expr, file_path) {
   check <- try(expr, silent = TRUE)
   if (inherits(check, "try-error")) {
-    message <- attr(check, "condition")$message
-    parent_msg <- attr(check, "condition")$parent$message
-    if (is.character(parent_msg)) {
-      parent_msg <- paste(parent_msg, collapse = " --> ")
-      msg <- paste(message, parent_msg,  sep = " --> ")
-    } else {
-      msg <- message
-    }
-    msg <- clean_msg(msg)
+    msg <- as.character(check) %>%
+      cli::ansi_strip() %>%
+      clean_msg()
 
     return(
       capture_exec_error(