+++ tests/testthat/test-parse_file_name.R
@@ -15,6 +15,21 @@ test_that("parse_file_name works", {
     )
   )
 
+
+  expect_snapshot(
+    parse_file_name(
+      "hub-baseline.yml",
+      file_type = "model_metadata"
+    )
+  )
+  expect_snapshot(
+    parse_file_name(
+      "hubBaseline.yml",
+      file_type = "model_metadata"
+    ),
+    error = TRUE
+  )
+
   file_path <- "model-output/team1-goodmodel/2022-10-08-team1-goodmodel.csv"
   file_meta <- parse_file_name(file_path)
   expect_equal(
@@ -31,6 +46,8 @@ test_that("parse_file_name works", {
             sep = "-") %>%
           paste(file_meta$ext, sep = ".")
   )
+
+
 })
 
 test_that("parse_file_name fails correctly", {