+++ tests/testthat/test-read_model_out_file.R
@@ -11,8 +11,7 @@ test_that("read_model_out_file works", {
     str(
       read_model_out_file(
         file_path = "team1-goodmodel/2022-10-08-team1-goodmodel.csv",
-        hub_path = system.file("testhubs/simple", package = "hubValidations"),
-        use_hub_schema = TRUE
+        hub_path = system.file("testhubs/simple", package = "hubValidations")
       )
     )
   )